requirements.txt
apispec==6.3.0
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.1
azure-eventhub==5.11.6
azure-identity==1.14.1
azure-keyvault-secrets==4.7.0
blinker==1.6.2
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coverage==7.1.0
cryptography==42.0.5
decorator==5.1.1
distlib==0.3.7
et-xmlfile==1.1.0
excel-util==0.0.4
executing==2.0.1
filelock==3.12.4
fire==0.6.0
Flask==2.3.2
flask-apispec==0.11.4
Flask-Cors==4.0.0
future==1.0.0
google-auth==2.23.0
greenlet==2.0.2
idna==3.6
iniconfig==2.0.0
ipython==8.24.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
kubernetes==27.2.0
log4python==1.0.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdc==1.2.1
msal==1.27.0
msal-extensions==1.1.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
pandas-util==0.0.1
parso==0.8.4
platformdirs==3.10.0
pluggy==1.3.0
portalocker==2.8.2
prompt-toolkit==3.0.43
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.3
tabulate==0.9.0
termcolor==2.4.0
toml==0.10.2
traitlets==5.14.3
types-toml==0.10.8.7
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.24.5
wcwidth==0.2.13
webargs==8.3.0
websocket-client==1.6.3
Werkzeug==2.3.7
